// intranet/src/app/core/helios-api-client/model/timelineMembreDTO.ts
/**
 * Helios API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TimelineMembreType } from './timelineMembreType';


export interface TimelineMembreDTO { 
    id: number;
    type: TimelineMembreType;
    date: string;
    commentaire: string | null;
}