// intranet/src/app/core/helios-api-client/model/timelineMembreType.ts
/**
 * Helios API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TimelineMembreTypes } from './timelineMembreTypes';


export interface TimelineMembreType { 
    id?: number;
    description?: string | null;
    code?: TimelineMembreTypes;
    creation?: string;
    modification?: string | null;
}
export namespace TimelineMembreType {
}